pokemon_cli/pyproject.toml
[project]
name = "pokemon-tcgp-cli"
version = "0.1.0"
description = "AI Analyst CLI for Pokemon TCGP"
requires-python = ">=3.12"
dependencies = [
    "boring-semantic-layer[mcp]>=0.3.7",
    "google-genai>=1.63.0",
    "ibis-framework[duckdb]>=12.0.0",
    "mcp>=1.26.0",
    "ollama>=0.4.7",
    "pokemon-tcgp-semantic-layer",
    "rich>=13.9.4",
    "tenacity>=9.1.2",
]

[tool.setuptools]
packages = ["pokemon_cli"]

[tool.uv.sources]
pokemon-tcgp-semantic-layer = { workspace = true }

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
asyncio_mode = "auto"
